@@ -1,9 +1,9 @@
 <?php
 
-return array(
-  'title' => 'Titre',
-  'body' => 'Body',
-  'add' => 'Add',
-  'delete' => 'Delete',
-  'edit' => 'Edit',
-);
+return [
+    'title' => 'Titre',
+    'body' => 'Body',
+    'add' => 'Add',
+    'delete' => 'Delete',
+    'edit' => 'Edit',
+];
