@@ -1,9 +1,9 @@
 <?php
 
-return array(
-  'title' => 'Titre',
-  'body' => 'Body',
-  'add' => 'Ajouter',
-  'delete' => 'Supprimer',
-  'edit' => 'Modifier',
-);
+return [
+    'title' => 'Titre',
+    'body' => 'Body',
+    'add' => 'Ajouter',
+    'delete' => 'Supprimer',
+    'edit' => 'Modifier',
+];
